Remove events logic from zc_common
